LayoutTests/imported/w3c/web-platform-tests/css/css-values/tree-counting/sibling-index-keyframe-value-dynamic-expected.txt

You should see a green square below.


FAIL Initially, the sibling-index() is 3 for #target assert_equals: expected "3" but got "auto"
FAIL Removing a preceding sibling of #target reduces the sibling-index() assert_equals: expected "2" but got "auto"
